update ng (no-verify)
