--- logprep/processor/calculator/fourFn.py
@@ -1,2 +1,0 @@
-# TODO: remove when ng is implemented
-class BNF: ...
